Fixed TMOAuth crash on iOS 6 and earlier. Fixes tumblr/TMTumblrSDK#56
iOS 7 introduced a new API for base64-encoding, -[NSData base64EncodedStringWithOptions:]. This
API was not available in iOS 6 and earlier, but another API, -[NSData base64Encoding], was. It
was deprecated in iOS 7. So this commit changes TMOAuth to use the new API if available, and
fallback to the deprecated API if it is not available.